+++ sphinxcontrib/spelling/directive.py
@@ -16,9 +16,8 @@ def add_good_words_to_document(env, docname, good_words):
     # Initialize the per-document good words list
     if not hasattr(env, 'spelling_document_words'):
         env.spelling_document_words = collections.defaultdict(list)
-    logger.info('Extending local dictionary for %s with %s',
-                env.docname, good_words)
-    print(env.docname, good_words)
+    logger.debug('Extending local dictionary for %s with %s',
+                 env.docname, good_words)
     env.spelling_document_words[env.docname].extend(good_words)